- done with home Overview card details

+++ app/src/main/java/com/moneymanager/repo/TAccounts.java
@@ -136,7 +136,7 @@ public class TAccounts implements IAccount {
 	public double getSumOfBalanceOfAllAccounts() {
 		final Cursor c = dbHelper.select(q_SUM_BALANCE_FROM_ALL_ACCOUNTS(), null);
 		if (c.moveToNext()) {
-			return c.getInt(c.getColumnIndex(BALANCE));
+			return c.getDouble(c.getColumnIndex(BALANCE));
 		} else {
 			return -1;
 		}
@@ -146,7 +146,7 @@ public class TAccounts implements IAccount {
 	public double getSumOfBalanceOfAccount(int selectedAccount) {
 		final Cursor c = dbHelper.select(q_SUM_BALANCE_OF_ACCOUNT(selectedAccount), null);
 		if (c.moveToNext()) {
-			return c.getInt(c.getColumnIndex(BALANCE));
+			return c.getDouble(c.getColumnIndex(BALANCE));
 		} else {
 			return -1;
 		}
